Info panel: add weight
With this option Hugo now sorts the *.md files depending on its weight
that you can set with the YAML option "weight:".  The lower the value
the higher the ranking, thus putting it more at the top.

This functionality is used in order to structure the links on the info
panel.  This method might seem redundant (you could just hard code them
into the HTML), but I'm tying to be future proof.

As I am looking forward to set the options from Org Mode which would
gave this feature a good reason to be implemented as such.

Being modular!