Dotemacs: display last git commit at top
At the top of the document there is now a new blog which displays the
last git commit message of the “dotemacs” repository.

This merely is an alpha version and I might move it somewhere else, but
I think the feature will be preserved.

Actually I might add it for this repository too, under the “About”
section.