Sync: add flag to hugo command => minify
This flag will generate a public/ directory with its output be
minified.  Someone probably goes on a rant of how minified stuff sucks,
but it is an easy way of saving some data.

Besides the browser renders it for you.  And if one wants to read the
source of this website, they can-- and should-- do it in this repository
and not in the dev-tools' sources tab.