Style: code overflow scroll => auto
When the content of the code block would overflow on the X-axis, then
let it be scollable.  If the content does not overflow show no scroll
bar.  With the prior option it would still show a scroll bar, which
wasn't bad on a chromium-based browser, but on pale moon it was.